+++ python/verisocks/__init__.py
@@ -1,4 +1,4 @@
 """Verisocks Python client API
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"