+++ python/verisocks/utils.py
@@ -29,19 +29,59 @@ def _format_path(cwd, path):
     return os.path.abspath(os.path.join(cwd, path))
 
 
+def setup_sim_run(elab_cmd, sim_cmd, capture_output=True):
+    """Run simulation setup commands.
+
+    This command is e.g. used by :py:meth:`setup_sim` with elaboration and
+    simulation commands formatted according to the provided arguments.
+
+    Args:
+        elab_cmd (list): Elaboration command. It has to be provided as a list
+            of command and arguments (see subprocess documentation). If None,
+            this step is skipped.
+        sim_cmd (list): Simulation command. It has to be provided as a list
+            of command and arguments (see subprocess documentation). This
+            command is mandatory and cannot be None.
+        capture_output (bool): Defines if stdout and stderr output
+            are "captured" (i.e. not visible).
+
+    Returns:
+        subprocess.Popen
+    """
+
+    if elab_cmd:
+        subprocess.check_call(elab_cmd)
+
+    if sim_cmd is None:
+        raise ValueError("Simulation command and arguments is mandatory")
+
+    if capture_output:
+        pop = subprocess.Popen(
+            sim_cmd,
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL
+        )
+    else:
+        pop = subprocess.Popen(sim_cmd)
+
+    logging.info(f"Launched Icarus with PID {pop.pid}")
+    return pop
+
+
 def setup_sim(vpi_libpath, *src_files, cwd=".", vvp_filepath=None,
               vvp_logpath="vvp.log", ivl_exec=None, ivl_args=None,
               vvp_exec=None, vvp_args=None, vvp_postargs=None,
               capture_output=True):
     """Set up Icarus simulation by elaborating the design with :code:`iverilog`
-    and launching the simulation with :code:`vvp`.
+    and launching the simulation with :code:`vvp`. Uses
+    :py:meth:`setup_sim_run` to run the concatenated commands and arguments.
 
     Args:
-        cwd (str): Reference path to be used for all paths provided as relative
-            paths.
         vpi_libpath (str): Path to the compiled Verisocks VPI library.
         src_files (str): Paths to all (verilog) source files to use for the
             simulation. All files have to be added as separate arguments.
+        cwd (str): Reference path to be used for all paths provided as relative
+            paths.
         vvp_filepath (str): Path to the elaborated VVP file (iverilog output).
             If None (default), "sim.vvp" will be used.
         vvp_logpath (str): Path to a simulation logfile. Default="vvp.log". If
@@ -95,7 +135,6 @@ def setup_sim(vpi_libpath, *src_files, cwd=".", vvp_filepath=None,
         *ivl_args,
         *src_file_paths
     ]
-    subprocess.check_call(ivl_cmd)
 
     # Simulation with vvp
     if vvp_exec:
@@ -121,14 +160,5 @@ def setup_sim(vpi_libpath, *src_files, cwd=".", vvp_filepath=None,
         *vvp_postargs
     ]
 
-    if capture_output:
-        pop = subprocess.Popen(
-            vvp_cmd,
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL
-        )
-    else:
-        pop = subprocess.Popen(vvp_cmd)
-
-    logging.info(f"Launched Icarus with PID {pop.pid}")
+    pop = setup_sim_run(ivl_cmd, vvp_cmd, capture_output)
     return pop