+++ docs/src/release_notes.rst
@@ -8,6 +8,20 @@ Releases of documentation and code are using the same version numbers. The
 version numbering system follows the `semantic versioning
 <https://semver.org/>`_ principles.
 
+1.2.0 - 2024-03-16
+******************
+
+* Modified :py:class:`Verisocks<verisocks.verisocks.Verisocks>` constructor and
+  :py:meth:`Verisocks.connect() <verisocks.verisocks.Verisocks.connect>` method
+  to include arguments for multiple, delayed connection trials. Examples and
+  test have been simplified accordingly.
+* Added correct management of system call interrupts while waiting on client
+  connection in the server code (see
+  https://www.gnu.org/software/libc/manual/html_node/Interrupted-Primitives.html
+  for details).
+* Added section :ref:`sec_alternative_simulators`.
+* Added method :py:meth:`verisocks.utils.setup_sim_run` to simplify foreseen
+  support for alternative simulators.
 
 1.1.0 - 2024-02-07
 ******************